Backend/requirements.txt
fastapi
uvicorn
python-dotenv
transformers
torch
nltk
scikit-learn
openai
joblib
